requirements.txt
requests
discord.py[voice]
litellm[proxy]
orjson
proxy
websocket-client
python-socketio[client]